+++ deepinv/tests/test_datasets.py
@@ -3,14 +3,7 @@ import shutil
 import PIL
 import pytest
 
-from deepinv.datasets import (
-    DIV2K,
-    Urban100HR,
-    Set14HR,
-    CBSD68,
-    LsdirHR,
-    FMD,
-)
+from deepinv.datasets import DIV2K, Urban100HR, Set14HR, CBSD68, LsdirHR, FMD, Kohler
 
 
 @pytest.fixture
@@ -125,6 +118,50 @@ def test_load_cbsd68_dataset(download_cbsd68):
 
 
 @pytest.fixture
+def download_Kohler():
+    """Download the Köhler dataset before a test and remove it after completion."""
+    root = "Kohler"
+    Kohler.download(root)
+
+    # Return the control flow to the test function
+    yield root
+
+    # Clean up the created directory
+    shutil.rmtree(root)
+
+
+@pytest.mark.skip(reason="Downloading Kohler dataset is unreliable for testing.")
+def test_load_Kohler_dataset(download_Kohler):
+    """Check that the Köhler dataset contains 48 PIL images."""
+    root = download_Kohler
+
+    dataset = Kohler(
+        root=root, frames="middle", ordering="printout_first", download=False
+    )
+    x1, y1 = dataset.get_item(1, 1, "middle")
+    x2, y2 = dataset[0]
+
+    assert (
+        len(dataset) == 48
+    ), f"The dataset should have been of len 48, instead got {len(dataset)}."
+
+    assert (
+        type(x1) == PIL.PngImagePlugin.PngImageFile
+    ), "The sharp frame is unexpectedly not a PIL image."
+
+    assert (
+        type(y1) == PIL.PngImagePlugin.PngImageFile
+    ), "The blurry frame is unexpectedly not a PIL image."
+
+    assert (
+        type(x2) == PIL.PngImagePlugin.PngImageFile
+    ), "The sharp frame is unexpectedly not a PIL image."
+
+    assert (
+        type(y2) == PIL.PngImagePlugin.PngImageFile
+    ), "The blurry frame is unexpectedly not a PIL image."
+
+
 def download_lsdir():
     """Downloads dataset for tests and removes it after test executions."""
     tmp_data_dir = "LSDIR"