+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ New Features
 - Multi-coil MRI, 3D MRI, MRI Mixin (:gh:`287` by `Andrew Wang`_, Brett Levac)
 - Add Metric baseclass, unified params (for complex, norm, reduce), typing, tests, L1L2 metric, QNR metric, metrics docs section, Metric functional wrapper (:gh:`309`, :gh:`343` by `Andrew Wang`_)
 - generate_dataset features: complex numbers, save/load physics_generator params, overwrite bool (:gh:`324`, :gh:`352` by `Andrew Wang`_)
+- Add the Köhler dataset (:gh:`271` by `Jérémy Scanvic`_)
 
 Fixed
 ^^^^^